requirements.txt
selenium==4.21.0
pytest~=8.3.2
allure-pytest==2.13.5
python-dateutil~=2.9.0.post0